+++ src/apis/MenuController.php
@@ -67,16 +67,24 @@ class MenuController extends RestController
 
         $log = [];
         foreach ($accessList as $access) {
-            $data = (new \yii\db\Query())->select(['timestamp_create', 'user_id', 'admin_ngrest_log.id', 'is_update', 'is_insert', 'admin_user.firstname', 'admin_user.lastname'])->from('admin_ngrest_log')->leftJoin('admin_user', 'admin_ngrest_log.user_id = admin_user.id')->orderBy('timestamp_create DESC')->limit(30)->where('api=:api and user_id!=0', [':api' => $access['permssionApiEndpoint']])->all();
+            $data = (new \yii\db\Query())->select(['timestamp_create', 'user_id', 'admin_ngrest_log.id', 'is_update', 'is_delete', 'is_insert', 'admin_user.firstname', 'admin_user.lastname'])->from('admin_ngrest_log')->leftJoin('admin_user', 'admin_ngrest_log.user_id = admin_user.id')->orderBy('timestamp_create DESC')->limit(30)->where('api=:api and user_id!=0', [':api' => $access['permssionApiEndpoint']])->all();
             foreach ($data as $row) {
                 $date = mktime(0, 0, 0, date('n', $row['timestamp_create']), date('j', $row['timestamp_create']), date('Y', $row['timestamp_create']));
+                if ($row['is_update']) {
+                    $message = Module::t('dashboard_log_message_edit', ['container' => $access['alias']]);
+                } elseif ($row['is_insert']) {
+                    $message = Module::t('dashboard_log_message_add', ['container' => $access['alias']]);
+                } elseif ($row['is_delete']) {
+                    $message = Module::t('dashboard_log_message_delete', ['container' => $access['alias']]);
+                }
                 $log[$date][] = [
                     'name' => $row['firstname'].' '.$row['lastname'],
                     'is_update' => $row['is_update'],
                     'is_insert' => $row['is_insert'],
+                    'is_delete' => $row['is_delete'],
                     'timestamp' => $row['timestamp_create'],
                     'alias' => $access['alias'],
-                    'message' => ($row['is_update']) ? Module::t('dashboard_log_message_edit', ['container' => $access['alias']]) : Module::t('dashboard_log_message_add', ['container' => $access['alias']]),
+                    'message' => $message,
                     'icon' => $access['icon'],
                 ];
             }