+++ src/views/template/default.php
@@ -50,6 +50,7 @@ use luya\admin\Module;
                     <div class="timeline-item timeline-item-center-point" ng-repeat="(key, log) in item.items">
                         <i class="material-icons" ng-if="log.is_insert == 1">note_add</i>
                         <i class="material-icons" ng-if="log.is_update == 1">create</i>
+                        <i class="material-icons" ng-if="log.is_delete == 1">delete</i>
                         <small class="pr-4"><i>{{log.timestamp * 1000 | date:"HH:mm"}}</i></small>
                         <small class="pr-4">{{ log.name }}</small>
                         <small><span compile-html ng-bind-html="log.message | trustAsUnsafe"></span></small>